Corrected the cost function
Previously the cost was increasing monotonically with every iteration, but this was due to the error in compute_cost function. Y and AL werent of the same dimension which caused an error in element wise multiplication (the operation still went through but it wasnt correct). the output layer activations AL were of the dimension (4,m) but Y was of the dimension (1,m). so had to one hot encode the true labels and then perform the element wise multiplication and it worked.